extract another network util
