+++ src/chart_overlay.py
@@ -5,20 +5,7 @@ from PIL import ImageDraw
 from src import price_humaniser
 from src.configuration.log_decorator import info_log
 from src.image_utils import bottom_right_text, bottom_left_text, top_right_text, border, rotated_center_right_text, DrawText, TextBlock, least_intrusive_position
-
-
-def get_ip():
-    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    s.settimeout(0)
-    try:
-        # doesn't even have to be reachable
-        s.connect(('10.255.255.255', 1))
-        ip = s.getsockname()[0]
-    except Exception:
-        ip = '127.0.0.1'
-    finally:
-        s.close()
-    return ip
+from src.network_utils import get_ip
 
 
 class ChartOverlay():
